Make Catalog the primary interface
