--- partiql-planner/src/main/kotlin/org/partiql/planner/metadata/Metadata.kt
@@ -1,12 +1,0 @@
-package org.partiql.planner.metadata
-
-/**
- * Top-level metadata interface for access to object descriptors.
- */
-public interface Metadata {
-
-    /**
-     * The root namespace.
-     */
-    public fun getNamespace(): Namespace
-}
